Fixing Airy Function implementation and other stuff
The Moyal/Airy boundary was changed, while the Edgeworth approximation was disabled. Further revisions will remove unnecessary functions.
The Airy approximation of Vavilov was corrected and introduced a new formula to determine the best boundaries for numerical integration.
Further testing are under private work before a new set of packages to be distributed.